Finished writing code for the ADC, and LCD
Tested the LCD, only functions by manually lowering the baudrate (i.e. the baudrate is running far too fast)

The ADC works. We hooked up the ten sensors and resolved angles successfully to within specifications with an IR LED.

Need to write some code to take a running average of the most recent values to smooth out the values.