Implemented logic for jumper and FT232.
Removing the jumper will put the system into polling mode.
Replacing the jumper will put it back to the idle state.

Powering the FT232 will put the system into PC communication mode.
Unplugging it *should* put it back to the idle state.
Need to implement a non-blocking getchar function or a timeout for this to happen.